Tear down and newly set up git-secret
Now using GnuPG 1.4 instead of 2.2 on Mac.

This is to make it work across machines: on Ubuntu, the most recent available
version is 1.4.20, and this doesn't work together with version 2.2.

Furthermore, git-secret has been tested for GnuPG 1.4.20
(http://git-secret.io/installation).

The reason is, I think, that GnuPG 2.2 produces a .gitsecret/keys/pubring.kbx
but GnuPG 1.4 expects a .gitsecret/keys/pubring.gpg.

So, when I tried to decrypt the files on an Ubuntu machine with GnuPG 1.4.20,
it looked like a git-secret repo with no added users (because the
expected .gitsecret/keys/pubring.gpg was missing).